TSP layer of Xen vendor command DeepQuote
Added TSP layer which implements Xen version of DeepQuote command.
It is used to provide deep attestation through all layers. It is an
extension of Tspi_TPM_Quote. It has 3 additional parameters
 * PCR composite param which can be used to create a selection of
physical TPM PCRs.
 * flags - a bitmask which is used by vTPM manager to add strength
to signature creation by including additional hashes in external data
 * vtpm_group_pubkey - the public key of the group, in which the vTPM
attached to virtual machine guest is running, is used as a signing key
by vTPM manager and it is required if internal validation is requested

Signed-off-by: Emil Condrea <emilcondrea@gmail.com>